Split benchmark dataset tests
